feat(cashflow): integrate asset name mapping and enhance cashflow page description
